git renamed affineGap lowMem (default now) and highMem
